Completed testing. Added suites for DAO classes in database and helper methods when needed.

+++ src/backend/Business.java
@@ -17,7 +17,7 @@ public class Business extends Account {
         details.saveBusinessType(businessType);
     }
 
-    public void saveAccessPrice(float accessPrice) {
+    public void saveAccessPrice(double accessPrice) {
         details.saveAccessPrice(accessPrice);
     }
 
@@ -66,5 +66,4 @@ public class Business extends Account {
         }
         return false;
     }
-
 }
